MNT: Add some metadata to the notebook plots
Interesting project :)

Was poking at the basic usage jupyter notebook and found that some of
the plots and code was not intuitive.  Added some in-line comments and
did some basic things to the plots like add labels to the x/y axes and
add a legend.  I was also curious to see what the cumulative savings
looked like, so I added that to one of the plots.

No hard feelings if you reject this PR. This was me just having some fun
with one of the notebooks in your repo.

Fwiw this was run with matplotlib 2.0 which has some neat new features
like understanding timeseries natively (and much better color defaults!)