Tests (Config): Move pytest.ini into tox.ini
Docs: https://docs.pytest.org/en/latest/customize.html

It's a matter of prefernece, but over time I've found having pytest
and tox settings to be in the same file to be the "Least Astonishment" way.

The concerns are still clearly separated in the file, and the file isn't too big.
It's just a bit more instant to see, and kind-of matches using 'py.test' or 'tox'
command alternately, as needed, but expecting each to load same config.