Remove pytest-runner from setup.py for Docker
This isn't a hard requirement but as it is, you need pytest (or pytest-runner)
wherever setup.py is running.

So to copy setup.py into  the Docker image and install that way,
you get stuck from the missing requirement.

There are three ways to address this:
* Set up the Dockerfile by installing the package from PyPI (etc) (so setup.py is not needed)
* Install pytest etc in the Docker image so this works out
* Remove pytest-runner test_requires so that you don't have to install pytest etc

I looove pytest but I opt for removing the need here,
so that we can have a tiny Docker image without extra stuff.

Open to doing the other options too, happy to discuss.