Add Dockerfile (meant for portable local usage)
This is meant to offer a way to try eemeter
without the hurdle of installing the Python scientific stack.
Mainly that is a hurdle for Windows users.

----

This is not meant for any kind of long-running process as it is.
Just short-lived CLI usage.

----

Note, I'm *not* assuming that Windows users typically have docker installed.
However these days it's trivially easy to install Docker For Windows or Docker For Mac.
Once they do that, then there's no more complexity about Scientific Computing Stack
having issues for installing from source etc.

this *does* complicate commandline usage a bit,
but a later enhancement could create scripts (whether bash-for-cygwin or PowerShell)
to get the usage back to simple usage, using the Docker container under the hood.

----

Another future enhancement note:

Could do a variant of this that even helps with development usage,
deriving from this image but copying in source and tests and running in container.
If there any contributors who have trouble setting up their local envs.

----

Usage example - assuming you've done a local `docker build` and named container 'eemeter'

    docker run eemeter --help
    docker run eemeter sample

    docker run --volume $(pwd)/eemeter/sample_data:/tmp/input_data eemeter analyze /tmp/input_data

    # or
    docker run -it --volume $(pwd)/eemeter/sample_data:/tmp/input_data eemeter analyze /tmp/input_data

(`-it` is often passed for a local CLI usage, but not strictly necessary in this case.
however it does make the output stream out faster which is a nicer experience.
plus it allows for later passing things in on 'stdin' if we want that.)