Add 'bandit' security static analysis to tox & Travis
Documented with some comments in tox.ini, which should explain I hope!

The official bandit docs are good too.

https://github.com/openstack/bandit

I separated into a 'baseline testing' that is stricter, but
more actionable as it can run locally;
and 'gatekeeping' that only fails-out on HIGH severity,
and that part runs in Travis to put the breaks on and
have us fix or mitigate the issue before merge.