test_noaa_client.py fixtures - no more Internet :)
Added a script to "capture" responses from NOAA and
help save those as a test suite fixture.

Then load in and mock one method in NOAAClient to load
from that previously saved file.

In spirit it's similar to the VCR.py library,
but that only works for HTTP/S, not FTP.

It might seem like a lot for just 2 test cases,
but now I think this opens up to add a variety of further
test cases if desired.

Bottom line: test_noaa_client.py doesn't talk to the Internet anymore!